Set isRestricted to true when query string has key
It should be possible to show a restricted MAM channel when the query
string parameters address and key are available. Because the
isRestricted global variable is only set to true when selection
restricted in the dropdown this didn't work any more. This commit fixes
that by setting isRestricted to true when an address and sidekey are
available in the query string.